style: move about page component to folder
