+++ lib/components/team-page/footprints.tsx
